switch to dds_serialize free function
