+++ dds/src/implementation/data_representation_builtin_endpoints/discovered_topic_data.rs
@@ -2,19 +2,21 @@ use crate::{
     builtin_topics::TopicBuiltinTopicData,
     implementation::parameter_list_serde::{
         parameter_list_deserializer::ParameterListDeserializer,
-        parameter_list_serializer::ParameterListSerializer,
     },
     infrastructure::error::DdsResult,
     topic_definition::type_support::{
-        DdsDeserialize, DdsSerialize, DdsSerializedKey, DdsType, RepresentationType, PL_CDR_LE,
+        DdsDeserialize, DdsSerializedKey, DdsType, RepresentationType, PL_CDR_LE, RepresentationFormat,
     },
 };
 
 pub const DCPS_TOPIC: &str = "DCPSTopic";
 
-#[derive(Debug, PartialEq, Eq)]
+#[derive(Debug, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
 pub struct DiscoveredTopicData {
     topic_builtin_topic_data: TopicBuiltinTopicData,
+}
+impl RepresentationFormat for DiscoveredTopicData {
+    const REPRESENTATION_IDENTIFIER: RepresentationType = PL_CDR_LE;
 }
 
 impl DiscoveredTopicData {
@@ -50,18 +52,6 @@ impl DdsType for DiscoveredTopicData {
     }
 }
 
-impl DdsSerialize for DiscoveredTopicData {
-    const REPRESENTATION_IDENTIFIER: RepresentationType = PL_CDR_LE;
-
-    fn dds_serialize_parameter_list<W: std::io::Write>(
-        &self,
-        serializer: &mut ParameterListSerializer<W>,
-    ) -> DdsResult<()> {
-        self.topic_builtin_topic_data
-            .dds_serialize_parameter_list(serializer)
-    }
-}
-
 impl<'de> DdsDeserialize<'de> for DiscoveredTopicData {
     fn dds_deserialize_parameter_list<E: byteorder::ByteOrder>(
         deserializer: &mut ParameterListDeserializer<'de, E>,
@@ -75,6 +65,7 @@ impl<'de> DdsDeserialize<'de> for DiscoveredTopicData {
 #[cfg(test)]
 mod tests {
     use crate::builtin_topics::BuiltInTopicKey;
+    use crate::implementation::parameter_list_serde::serde_parameter_list_serializer::dds_serialize;
     use crate::infrastructure::qos_policy::{
         DeadlineQosPolicy, DestinationOrderQosPolicy, DurabilityQosPolicy, HistoryQosPolicy,
         LatencyBudgetQosPolicy, LifespanQosPolicy, LivelinessQosPolicy, OwnershipQosPolicy,
@@ -83,12 +74,6 @@ mod tests {
     };
 
     use super::*;
-
-    fn to_bytes_le<S: DdsSerialize>(value: &S) -> Vec<u8> {
-        let mut writer = Vec::<u8>::new();
-        value.dds_serialize(&mut writer).unwrap();
-        writer
-    }
 
     #[test]
     fn serialize_all_default() {
@@ -129,7 +114,8 @@ mod tests {
             b'c', b'd', 0, 0x00, // DomainTag: string + padding (1 byte)
             0x01, 0x00, 0x00, 0x00, // PID_SENTINEL, length
         ];
-        assert_eq!(to_bytes_le(&data), expected);
+        let result = dds_serialize(&data).unwrap();
+        assert_eq!(result, expected);
     }
 
     #[test]
