+++ dds/src/implementation/parameter_list_serde/mod.rs
@@ -1,5 +1,5 @@
 pub mod parameter_list_deserializer;
 pub mod parameter_list_serializer;
-// pub mod parameter;
+pub mod parameter;
 pub mod serde_parameter_list_deserializer;
 pub mod serde_parameter_list_serializer;
