+++ dds/src/implementation/parameter_list_serde/parameter.rs
@@ -1,210 +1,280 @@
-use std::marker::PhantomData;
+use std::{self, marker::PhantomData};
 
-use serde::ser::SerializeStruct;
+use serde::de::{self};
+use serde::ser::SerializeTuple;
 
-type Pid = u16;
-pub struct Parameter<T> {
-    pid: Pid,
-    value: T,
-}
+pub type RepresentationType = [u8; 2];
+pub type RepresentationOptions = [u8; 2];
+pub const CDR_BE: RepresentationType = [0x00, 0x00];
+pub const CDR_LE: RepresentationType = [0x00, 0x01];
+pub const PL_CDR_BE: RepresentationType = [0x00, 0x02];
+pub const PL_CDR_LE: RepresentationType = [0x00, 0x03];
+pub const REPRESENTATION_OPTIONS: RepresentationOptions = [0x00, 0x00];
 
-impl<T> Parameter<T> {
-    pub fn new(pid: Pid, value: T) -> Self {
-        Self { pid, value }
+pub const PID_SENTINEL: u16 = 1;
+
+#[derive(Debug, PartialEq, Eq, Clone)]
+pub struct Parameter<const PID: u16, T>(pub T);
+
+#[derive(Debug, PartialEq, Eq, Clone)]
+pub struct ParameterWithDefault<const PID: u16, T>(pub T);
+
+
+#[derive(Debug, PartialEq, Eq, Clone)]
+pub struct ParameterVector<const PID: u16, T>(pub Vec<T>);
+
+
+impl<const PID: u16, T> From<T> for Parameter<PID, T> {
+    fn from(v: T) -> Self {
+        Parameter(v)
     }
 }
 
-impl<T: serde::Serialize> serde::Serialize for Parameter<T> {
+impl<const PID: u16, T> From<T> for ParameterWithDefault<PID, T> {
+    fn from(v: T) -> Self {
+        ParameterWithDefault(v)
+    }
+}
+
+impl<const PID: u16, T> From<Vec<T>> for ParameterVector<PID, T> {
+    fn from(v: Vec<T>) -> Self {
+        ParameterVector(v)
+    }
+}
+
+
+impl<const PID: u16, T> serde::Serialize for Parameter<PID, T>
+where
+    T: serde::Serialize,
+{
     fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
     where
         S: serde::Serializer,
     {
-        let length_without_padding = cdr::size::calc_serialized_data_size(&self.value) as i16;
+        let length_without_padding = cdr::size::calc_serialized_data_size(&self.0) as i16;
         let padding_length = (4 - length_without_padding) & 3;
         let length = length_without_padding + padding_length;
 
-        let mut s = serializer.serialize_struct("parameter", 4)?;
-        s.serialize_field("pid", &self.pid)?;
-        s.serialize_field("length", &length)?;
-        s.serialize_field("value", &self.value)?;
+        let mut s = serializer.serialize_tuple(4)?;
+        s.serialize_element(&PID)?;
+        s.serialize_element(&length)?;
+        s.serialize_element(&self.0)?;
         match padding_length {
-            1 => s.serialize_field("padding", &[0u8; 1])?,
-            2 => s.serialize_field("padding", &[0u8; 2])?,
-            3 => s.serialize_field("padding", &[0u8; 3])?,
-            _ => s.serialize_field("padding", &[0u8; 0])?,
+            1 => s.serialize_element(&[0u8; 1])?,
+            2 => s.serialize_element(&[0u8; 2])?,
+            3 => s.serialize_element(&[0u8; 3])?,
+            _ => s.serialize_element(&[0u8; 0])?,
         }
         s.end()
     }
 }
 
-struct ParameterVisitor<T> {
-    phantom: PhantomData<T>,
-}
-
-impl<T> ParameterVisitor<T> {
-    fn new() -> Self {
-        Self {
-            phantom: PhantomData,
-        }
-    }
-}
-
-impl<'de, T: serde::Deserialize<'de> + serde::Serialize> serde::de::Visitor<'de>
-    for ParameterVisitor<T>
+impl<const PID: u16, T> serde::Serialize for ParameterWithDefault<PID, T>
+where
+    T: serde::Serialize + Default + PartialEq,
 {
-    type Value = Parameter<T>;
-
-    fn expecting(&self, formatter: &mut std::fmt::Formatter) -> std::fmt::Result {
-        formatter.write_str("Parameter")
-    }
-
-    fn visit_seq<V>(self, mut seq: V) -> Result<Self::Value, V::Error>
-    where
-        V: serde::de::SeqAccess<'de>,
-    {
-        let pid = seq
-            .next_element()?
-            .ok_or_else(|| serde::de::Error::invalid_length(0, &self))?;
-        let length: i16 = seq
-            .next_element()?
-            .ok_or_else(|| serde::de::Error::invalid_length(1, &self))?;
-        let value = seq
-            .next_element()?
-            .ok_or_else(|| serde::de::Error::invalid_length(2, &self))?;
-
-        let length_without_padding = cdr::size::calc_serialized_data_size(&value) as i16;
-        let padding_length = length - length_without_padding;
-        match padding_length {
-            1 => {
-                let _padding: [u8; 1] = seq
-                    .next_element()?
-                    .ok_or_else(|| serde::de::Error::invalid_length(3, &self))?;
-            }
-            2 => {
-                let _padding: [u8; 2] = seq
-                    .next_element()?
-                    .ok_or_else(|| serde::de::Error::invalid_length(3, &self))?;
-            }
-            3 => {
-                let _padding: [u8; 3] = seq
-                    .next_element()?
-                    .ok_or_else(|| serde::de::Error::invalid_length(3, &self))?;
-            }
-            _ => (),
-        };
-        Ok(Parameter::new(pid, value))
-    }
-}
-
-impl<'de, T: serde::Deserialize<'de> + serde::Serialize> serde::Deserialize<'de> for Parameter<T> {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        deserializer.deserialize_struct(
-            "parameter",
-            &["pid", "length", "value", "padding"],
-            ParameterVisitor::new(),
-        )
-    }
-}
-
-impl<'de, T: serde::Deserialize<'de>> serde::Deserialize<'de> for ParameterWithDefault<T> {
-    fn deserialize<D>(_deserializer: D) -> Result<Self, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        todo!()
-    }
-}
-
-pub struct ParameterWithDefault<T> {
-    parameter: Parameter<T>,
-    default: T,
-}
-
-impl<T> ParameterWithDefault<T> {
-    pub fn new(pid: Pid, value: T) -> Self
-    where
-        T: Default,
-    {
-        Self {
-            parameter: Parameter { pid, value },
-            default: T::default(),
-        }
-    }
-}
-impl<T: serde::Serialize + PartialEq> serde::Serialize for ParameterWithDefault<T> {
     fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
     where
         S: serde::Serializer,
     {
-        if self.parameter.value == self.default {
-            serializer.serialize_struct("", 0)?.end()
+        if self.0 == T::default() {
+            serializer.serialize_tuple(0)?.end()
         } else {
-            self.parameter.serialize(serializer)
+            let length_without_padding = cdr::size::calc_serialized_data_size(&self.0) as i16;
+            let padding_length = (4 - length_without_padding) & 3;
+            let length = length_without_padding + padding_length;
+
+            let mut s = serializer.serialize_tuple(4)?;
+            s.serialize_element(&PID)?;
+            s.serialize_element(&length)?;
+            s.serialize_element(&self.0)?;
+            match padding_length {
+                1 => s.serialize_element(&[0u8; 1])?,
+                2 => s.serialize_element(&[0u8; 2])?,
+                3 => s.serialize_element(&[0u8; 3])?,
+                _ => s.serialize_element(&[0u8; 0])?,
+            }
+            s.end()
         }
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use serde::Serialize;
-
-    use super::*;
-    use crate::{
-        implementation::data_representation_builtin_endpoints::parameter_id_values::PID_SENTINEL,
-        topic_definition::type_support::{PL_CDR_LE, REPRESENTATION_OPTIONS},
-    };
-
-    #[derive(serde::Serialize, serde::Deserialize)]
-    struct Inner {
-        id: Parameter<u8>,
-    }
-
-    #[derive(serde::Serialize, serde::Deserialize)]
-    struct Outer {
-        inner: Inner,
-        qos: ParameterWithDefault<Qos>,
-        text: Parameter<String>,
-    }
-
-    #[derive(serde::Serialize, serde::Deserialize, PartialEq)]
-    struct Qos {
-        is: bool,
-    }
-    impl Default for Qos {
-        fn default() -> Self {
-            Self { is: false }
+impl<const PID: u16, T> serde::Serialize for ParameterVector<PID, T>
+where
+    T: serde::Serialize,
+{
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: serde::Serializer,
+    {
+        if self.0.len() == 0 {
+            return serializer.serialize_tuple(0)?.end();
         }
+        let length_without_padding = cdr::size::calc_serialized_data_size(&self.0[0]) as i16;
+        let padding_length = (4 - length_without_padding) & 3;
+        let length = length_without_padding + padding_length;
+
+        let mut s = serializer.serialize_tuple(4)?;
+        for value in &self.0 {
+            s.serialize_element(&PID)?;
+            s.serialize_element(&length)?;
+            s.serialize_element(value)?;
+            match padding_length {
+                1 => s.serialize_element(&[0u8; 1])?,
+                2 => s.serialize_element(&[0u8; 2])?,
+                3 => s.serialize_element(&[0u8; 3])?,
+                _ => s.serialize_element(&[0u8; 0])?,
+            }
+        }
+        s.end()
     }
+}
 
-    #[test]
-    fn serialize_simple() {
-        let data = Outer {
-            inner: Inner {
-                id: Parameter::new(71, 21),
+
+
+
+impl<'de, const PID: u16, T> serde::Deserialize<'de> for Parameter<PID, T>
+where
+    T: serde::Deserialize<'de>,
+{
+    fn deserialize<D>(deserializer: D) -> std::result::Result<Self, D::Error>
+    where
+        D: serde::Deserializer<'de>,
+    {
+        struct Visitor<'de, const PID: u16, T>
+        where
+            T: serde::Deserialize<'de>,
+        {
+            marker: PhantomData<Parameter<PID, T>>,
+            lifetime: PhantomData<&'de ()>,
+        }
+        impl<'de, const PID: u16, T> serde::de::Visitor<'de> for Visitor<'de, PID, T>
+        where
+            T: serde::Deserialize<'de>,
+        {
+            type Value = Parameter<PID, T>;
+            fn expecting(&self, formatter: &mut std::fmt::Formatter) -> std::fmt::Result {
+                formatter.write_str("struct Parameter")
+            }
+
+            fn visit_map<A>(self, mut map: A) -> std::result::Result<Self::Value, A::Error>
+            where
+                A: de::MapAccess<'de>,
+            {
+                loop {
+                    if let Some(key) = map.next_key::<u16>()? {
+                        if key == PID {
+                            return Ok(Parameter(map.next_value()?));
+                        }
+                    }
+                }
+            }
+        }
+        deserializer.deserialize_newtype_struct(
+            "Parameter",
+            Visitor {
+                marker: PhantomData::<Parameter<PID, T>>,
+                lifetime: PhantomData,
             },
-            qos: ParameterWithDefault::new(72, Qos { is: false }),
-            text: Parameter::new(73, "ab".to_string()),
-        };
-        let mut result = vec![];
-        let mut serializer = cdr::ser::Serializer::<_, byteorder::LittleEndian>::new(&mut result);
-        PL_CDR_LE.serialize(&mut serializer).unwrap();
-        REPRESENTATION_OPTIONS.serialize(&mut serializer).unwrap();
-        data.serialize(&mut serializer).unwrap();
-        PID_SENTINEL.serialize(&mut serializer).unwrap();
-        [0_u8, 0].serialize(&mut serializer).unwrap();
+        )
+    }
+}
 
-        let expected = vec![
-            0x00, 0x03, 0x00, 0x00, // PL_CDR_LE | OPTIONS
-            71, 0x00, 4, 0, // id | Length (incl padding)
-            21, 0, 0, 0, // u8
-            73, 0x00, 8, 0x00, // text | Length  (incl padding)
-            3, 0x00, 0x00, 0x00, // String length (incl. termination)
-            b'a', b'b', 0, 0x00, // String + termination 0 + padding (1 byte)
-            0x01, 0x00, 0x00, 0x00, // PID_SENTINEL
-        ];
-        assert_eq!(result, expected)
+
+impl<'de, const PID: u16, T> serde::Deserialize<'de> for ParameterWithDefault<PID, T>
+where
+    T: serde::Deserialize<'de> + Default,
+{
+    fn deserialize<D>(deserializer: D) -> std::result::Result<Self, D::Error>
+    where
+        D: serde::Deserializer<'de>,
+    {
+        struct Visitor<'de, const PID: u16, T>
+        where
+            T: serde::Deserialize<'de>,
+        {
+            marker: PhantomData<ParameterWithDefault<PID, T>>,
+            lifetime: PhantomData<&'de ()>,
+        }
+        impl<'de, const PID: u16, T> serde::de::Visitor<'de> for Visitor<'de, PID, T>
+        where
+            T: serde::Deserialize<'de> + Default,
+        {
+            type Value = ParameterWithDefault<PID, T>;
+            fn expecting(&self, formatter: &mut std::fmt::Formatter) -> std::fmt::Result {
+                formatter.write_str("struct ParameterWithDefault")
+            }
+
+            fn visit_map<A>(self, mut map: A) -> std::result::Result<Self::Value, A::Error>
+            where
+                A: de::MapAccess<'de>,
+            {
+                while let Some(key) = map.next_key::<u16>()? {
+                    if key == PID {
+                        return Ok(ParameterWithDefault(map.next_value()?));
+                    } else if key == PID_SENTINEL {
+                        break;
+                    }
+                }
+                Ok(ParameterWithDefault(T::default()))
+            }
+        }
+        deserializer.deserialize_newtype_struct(
+            "ParameterWithDefault",
+            Visitor {
+                marker: PhantomData::<ParameterWithDefault<PID, T>>,
+                lifetime: PhantomData,
+            },
+        )
+    }
+}
+
+
+impl<'de, const PID: u16, T> serde::Deserialize<'de> for ParameterVector<PID, T>
+where
+    T: serde::Deserialize<'de>,
+{
+    fn deserialize<D>(deserializer: D) -> std::result::Result<Self, D::Error>
+    where
+        D: serde::Deserializer<'de>,
+    {
+        struct Visitor<'de, const PID: u16, T>
+        where
+            T: serde::Deserialize<'de>,
+        {
+            marker: PhantomData<ParameterVector<PID, T>>,
+            lifetime: PhantomData<&'de ()>,
+        }
+        impl<'de, const PID: u16, T> serde::de::Visitor<'de> for Visitor<'de, PID, T>
+        where
+            T: serde::Deserialize<'de>,
+        {
+            type Value = ParameterVector<PID, T>;
+            fn expecting(&self, formatter: &mut std::fmt::Formatter) -> std::fmt::Result {
+                formatter.write_str("struct ParameterVector")
+            }
+
+            fn visit_map<A>(self, mut map: A) -> std::result::Result<Self::Value, A::Error>
+            where
+                A: de::MapAccess<'de>,
+            {
+                let mut values = vec![];
+                while let Some(key) = map.next_key::<u16>()? {
+                    if key == PID {
+                        values.push(map.next_value()?);
+                    } else if key == PID_SENTINEL {
+                        break;
+                    }
+                }
+                Ok(ParameterVector(values))
+            }
+        }
+        deserializer.deserialize_newtype_struct(
+            "ParameterVector",
+            Visitor {
+                marker: PhantomData::<ParameterVector<PID, T>>,
+                lifetime: PhantomData,
+            },
+        )
     }
 }
