+++ dds/src/implementation/parameter_list_serde/serde_parameter_list_serializer.rs
@@ -3,11 +3,11 @@ use std::marker::PhantomData;
 use std::{self};
 
 use byteorder::{ByteOrder, WriteBytesExt};
-use serde::ser::SerializeTuple;
 
 use cdr::Error;
-
 use crate::topic_definition::type_support::RepresentationFormat;
+
+use super::parameter::{REPRESENTATION_OPTIONS, PID_SENTINEL};
 
 pub struct ParameterListSerializer<W, E> {
     ser: cdr::Serializer<W, E>,
@@ -352,135 +352,6 @@ impl<'a, W, E> serde::ser::SerializeStructVariant for Compound<'a, W, E> {
     }
 }
 
-#[derive(Debug, PartialEq, serde::Serialize)]
-struct XParameter<const PID: u16, T>(T);
-
-#[derive(Debug, PartialEq, Eq, Clone)]
-pub struct Parameter<const PID: u16, T>(pub T);
-
-impl<const PID: u16, T> From<T> for Parameter<PID, T> {
-    fn from(v: T) -> Self {
-        Parameter(v)
-    }
-}
-// impl<const PID: u16, T> Into<T> for Parameter<PID, T> {
-//     fn into(self) -> T {
-//         todo!()
-//     }
-// }
-
-impl<const PID: u16, T> serde::Serialize for Parameter<PID, T>
-where
-    T: serde::Serialize,
-{
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        let length_without_padding = cdr::size::calc_serialized_data_size(&self.0) as i16;
-        let padding_length = (4 - length_without_padding) & 3;
-        let length = length_without_padding + padding_length;
-
-        let mut s = serializer.serialize_tuple(4)?;
-        s.serialize_element(&PID)?;
-        s.serialize_element(&length)?;
-        s.serialize_element(&self.0)?;
-        match padding_length {
-            1 => s.serialize_element(&[0u8; 1])?,
-            2 => s.serialize_element(&[0u8; 2])?,
-            3 => s.serialize_element(&[0u8; 3])?,
-            _ => s.serialize_element(&[0u8; 0])?,
-        }
-        s.end()
-    }
-}
-
-#[derive(Debug, PartialEq, Eq, Clone)]
-pub struct ParameterWithDefault<const PID: u16, T>(pub T);
-impl<const PID: u16, T> From<T> for ParameterWithDefault<PID, T> {
-    fn from(v: T) -> Self {
-        ParameterWithDefault(v)
-    }
-}
-impl<const PID: u16, T> serde::Serialize for ParameterWithDefault<PID, T>
-where
-    T: serde::Serialize + Default + PartialEq,
-{
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        if self.0 == T::default() {
-            serializer.serialize_tuple(0)?.end()
-        } else {
-            let length_without_padding = cdr::size::calc_serialized_data_size(&self.0) as i16;
-            let padding_length = (4 - length_without_padding) & 3;
-            let length = length_without_padding + padding_length;
-
-            let mut s = serializer.serialize_tuple(4)?;
-            s.serialize_element(&PID)?;
-            s.serialize_element(&length)?;
-            s.serialize_element(&self.0)?;
-            match padding_length {
-                1 => s.serialize_element(&[0u8; 1])?,
-                2 => s.serialize_element(&[0u8; 2])?,
-                3 => s.serialize_element(&[0u8; 3])?,
-                _ => s.serialize_element(&[0u8; 0])?,
-            }
-            s.end()
-        }
-    }
-}
-
-
-#[derive(Debug, PartialEq, Eq, Clone)]
-pub struct ParameterVector<const PID: u16, T>(pub Vec<T>);
-impl<const PID: u16, T> From<Vec<T>> for ParameterVector<PID, T> {
-    fn from(v: Vec<T>) -> Self {
-        ParameterVector(v)
-    }
-}
-impl<const PID: u16, T> serde::Serialize for ParameterVector<PID, T>
-where
-    T: serde::Serialize + Default + PartialEq,
-{
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        if self.0.len() == 0 {
-            return serializer.serialize_tuple(0)?.end();
-        }
-        let length_without_padding = cdr::size::calc_serialized_data_size(&self.0[0]) as i16;
-        let padding_length = (4 - length_without_padding) & 3;
-        let length = length_without_padding + padding_length;
-
-        let mut s = serializer.serialize_tuple(4)?;
-        for value in &self.0 {
-            s.serialize_element(&PID)?;
-            s.serialize_element(&length)?;
-            s.serialize_element(value)?;
-            match padding_length {
-                1 => s.serialize_element(&[0u8; 1])?,
-                2 => s.serialize_element(&[0u8; 2])?,
-                3 => s.serialize_element(&[0u8; 3])?,
-                _ => s.serialize_element(&[0u8; 0])?,
-            }
-        }
-        s.end()
-    }
-}
-
-pub type RepresentationType = [u8; 2];
-pub type RepresentationOptions = [u8; 2];
-pub const CDR_BE: RepresentationType = [0x00, 0x00];
-pub const CDR_LE: RepresentationType = [0x00, 0x01];
-pub const PL_CDR_BE: RepresentationType = [0x00, 0x02];
-pub const PL_CDR_LE: RepresentationType = [0x00, 0x03];
-pub const REPRESENTATION_OPTIONS: RepresentationOptions = [0x00, 0x00];
-
-const PID_SENTINEL: u16 = 1;
-
 pub fn dds_serialize<T>(value: &T) -> Result<Vec<u8>, Error>
 where
     T: serde::Serialize + RepresentationFormat,
@@ -525,7 +396,7 @@ where
 
 #[cfg(test)]
 mod tests {
-    use crate::topic_definition::type_support::DdsSerde;
+    use crate::{topic_definition::type_support::{DdsSerde, RepresentationFormat}, implementation::parameter_list_serde::parameter::{Parameter, RepresentationType, PL_CDR_LE, ParameterVector, ParameterWithDefault}};
 
     use super::*;
